my_project/requirements/dev.txt
-r base.txt
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
django-debug-toolbar==4.2.0
